Refactor: Detangle GitHub rendering from extraction
- Extracted `DependencyGraphRenderer` interface with single GitHub implementation
- Abstract plugin allows renderer classname to be specified

This should make it possible to implement support for other output formats.
