+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/LegacyDependencyExtractor.kt
@@ -2,7 +2,11 @@ package org.gradle.github.dependencygraph.internal
 
 import org.gradle.initialization.EvaluateSettingsBuildOperationType
 
-class LegacyDependencyExtractor : DependencyExtractor() {
+class LegacyDependencyExtractor(private val rendererClassName: String) : DependencyExtractor() {
+    override fun getRendererClassName(): String {
+        return rendererClassName
+    }
+
     override fun extractSettings(details: EvaluateSettingsBuildOperationType.Details) {
         // Extraction fails for included builds on Gradle 5.x.
         // It's OK to ignore these events since we only care about the root build settings file at this stage.
