+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/github/GitHubDependencyGraphRenderer.kt
@@ -1,21 +1,35 @@
 package org.gradle.github.dependencygraph.internal.github
 
 import org.gradle.api.logging.Logging
+import org.gradle.github.dependencygraph.internal.DependencyGraphRenderer
 import org.gradle.github.dependencygraph.internal.github.json.GitHubRepositorySnapshot
 import org.gradle.github.dependencygraph.internal.model.BuildLayout
 import org.gradle.github.dependencygraph.internal.model.ResolvedConfiguration
+import org.gradle.github.dependencygraph.internal.util.*
 import java.io.File
 
-class GitHubDependencyGraphOutput(private val snapshotParams: GitHubSnapshotParams, private val outputDir: File) {
+class GitHubDependencyGraphRenderer() : DependencyGraphRenderer {
 
-    fun outputDependencyGraph(resolvedConfigurations: MutableList<ResolvedConfiguration>, buildLayout: BuildLayout) {
+    override fun outputDependencyGraph(
+        pluginParameters: PluginParameters,
+        buildLayout: BuildLayout,
+        resolvedConfigurations: MutableList<ResolvedConfiguration>,
+        outputDirectory: File
+    ) {
+        val snapshotParams = GitHubSnapshotParams(
+            pluginParameters.load(PARAM_JOB_CORRELATOR),
+            pluginParameters.load(PARAM_JOB_ID),
+            pluginParameters.load(PARAM_GITHUB_SHA),
+            pluginParameters.load(PARAM_GITHUB_REF)
+        )
+
         val gitHubRepositorySnapshotBuilder = GitHubRepositorySnapshotBuilder(snapshotParams)
         // Use the job correlator as the manifest name
         val manifestName = snapshotParams.dependencyGraphJobCorrelator
         val manifest = gitHubRepositorySnapshotBuilder.buildManifest(manifestName, resolvedConfigurations, buildLayout)
         val snapshot = gitHubRepositorySnapshotBuilder.buildSnapshot(manifest)
 
-        val outputFile = File(outputDir, "${snapshotParams.dependencyGraphJobCorrelator}.json")
+        val outputFile = File(outputDirectory, "${snapshotParams.dependencyGraphJobCorrelator}.json")
 
         writeDependencySnapshot(snapshot, outputFile)
     }
@@ -23,10 +37,10 @@ class GitHubDependencyGraphOutput(private val snapshotParams: GitHubSnapshotPara
     private fun writeDependencySnapshot(graph: GitHubRepositorySnapshot, manifestFile: File) {
         manifestFile.parentFile.mkdirs()
         manifestFile.writeText(JacksonJsonSerializer.serializeToJson(graph))
-        LOGGER.lifecycle("\nGitHubDependencyGraphPlugin: Wrote dependency snapshot to \n${manifestFile.canonicalPath}")
+        LOGGER.lifecycle("\nGitHubDependencyGraphRenderer: Wrote dependency snapshot to \n${manifestFile.canonicalPath}")
     }
 
     companion object {
-        private val LOGGER = Logging.getLogger(GitHubDependencyGraphOutput::class.java)
+        private val LOGGER = Logging.getLogger(GitHubDependencyGraphRenderer::class.java)
     }
 }
